Refactor: move `my-ideas` and `my-solutions` to user feature
Moved `my-ideas` and `my-solutions` components to the `user` feature directory for improved organization and maintainability. Updated imports across relevant files to reflect the new file paths.
